// Backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  PARTICIPANT: 'PARTICIPANT',
  GUEST: 'GUEST'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const EventStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  CANCELED: 'CANCELED'
} as const

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus]


export const ReservationStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  REFUSED: 'REFUSED',
  CANCELED: 'CANCELED'
} as const

export type ReservationStatus = (typeof ReservationStatus)[keyof typeof ReservationStatus]